refactor(Editor): Renaming files. Spreading editor components
