+++ src/components/Editor/Editor.css
@@ -1,8 +1,8 @@
 .container {
-  border: 1px solid #E4E6EC;
+  border: 1px solid #e4e6ec;
   border-radius: 23px;
   padding: 10px 19px 0 19px;
-  background-color: #FFFFFF;
+  background-color: #ffffff;
 }
 
 blockquote {
