+++ src/components/Editor/components/EditorLeaf.tsx
@@ -1,0 +1,29 @@
+import React from 'react';
+
+export const EditorLeaf = ({
+  attributes,
+  children,
+  leaf,
+}: {
+  attributes: any;
+  children: any;
+  leaf: any;
+}) => {
+  if (leaf.bold) {
+    children = <strong>{children}</strong>;
+  }
+
+  if (leaf.code) {
+    children = <code>{children}</code>;
+  }
+
+  if (leaf.italic) {
+    children = <em>{children}</em>;
+  }
+
+  if (leaf.underline) {
+    children = <u>{children}</u>;
+  }
+
+  return <span {...attributes}>{children}</span>;
+};
