+++ src/components/Editor/components/index.ts
@@ -1,0 +1,8 @@
+export { EditorBlockButton } from './EditorBlockButton';
+export { EditorElement } from './EditorElement';
+export { EditorIcon } from './EditorIcon';
+export { EditorLeaf } from './EditorLeaf';
+export { EditorMarkButton } from './EditorMarkButton';
+export { EditorMenu } from './EditorMenu';
+export { EditorStateButton } from './EditorStateButton';
+export { EditorToolbar } from './EditorToolbar';
