+++ src/components/Editor/editor-base-props.ts
@@ -1,0 +1,4 @@
+export interface BaseProps {
+  className: string;
+  [key: string]: unknown;
+}
