--- src/components/WYSIWYGEditor/EditorComponents.tsx
@@ -1,203 +1,0 @@
-/** @jsx jsx */
-/* eslint react/jsx-key: 0 */
-import React, { PropsWithChildren, Ref } from 'react';
-import { useSlate } from 'slate-react';
-import { jsx } from 'theme-ui';
-import {
-  isBlockActive,
-  isMarkActive,
-  toggleBlock,
-  toggleMark,
-} from './EditorHelpers';
-
-export interface BaseProps {
-  className: string;
-  [key: string]: unknown;
-}
-
-export const Button = React.forwardRef(
-  (
-    {
-      className,
-      active,
-      reversed,
-      ...props
-    }: PropsWithChildren<
-      {
-        active: boolean;
-        reversed: boolean;
-      } & BaseProps
-    >,
-    ref: Ref<HTMLSpanElement>,
-  ) => (
-    <span
-      {...props}
-      ref={ref}
-      className={className}
-      sx={{
-        cursor: 'pointer',
-        color: reversed
-          ? active
-            ? 'white'
-            : '#aaa'
-          : active
-          ? 'black'
-          : '#ccc',
-      }}
-    />
-  ),
-);
-
-export const Icon = React.forwardRef(
-  (
-    { className, ...props }: PropsWithChildren<BaseProps>,
-    ref: Ref<HTMLSpanElement>,
-  ) => (
-    <div>
-      <link
-        rel="stylesheet"
-        href="https://fonts.googleapis.com/icon?family=Material+Icons"
-      />
-      <span
-        {...props}
-        ref={ref}
-        className={`material-icons ${className ? className : ''}`}
-      />
-    </div>
-  ),
-);
-
-export const MarkButton = ({
-  format,
-  icon,
-}: {
-  format: string;
-  icon: string;
-}) => {
-  const editor = useSlate();
-  return (
-    <Button
-      active={isMarkActive(editor, format)}
-      onMouseDown={(event: any) => {
-        event.preventDefault();
-        toggleMark(editor, format);
-      }}
-    >
-      <Icon>{icon}</Icon>
-    </Button>
-  );
-};
-
-export const BlockButton = ({
-  format,
-  icon,
-}: {
-  format: string;
-  icon: string;
-}) => {
-  const editor = useSlate();
-  return (
-    <Button
-      active={isBlockActive(editor, format)}
-      onMouseDown={(event: any) => {
-        event.preventDefault();
-        toggleBlock(editor, format);
-      }}
-    >
-      <Icon>{icon}</Icon>
-    </Button>
-  );
-};
-
-export const Menu = React.forwardRef(
-  (
-    { className, ...props }: PropsWithChildren<BaseProps>,
-    ref: Ref<HTMLDivElement>,
-  ) => (
-    <div
-      {...props}
-      ref={ref}
-      className={className}
-      sx={{
-        '& > *': {
-          display: 'inline-block',
-        },
-        '& > * + *': {
-          marginLeft: '15px',
-        },
-      }}
-    />
-  ),
-);
-
-export const Toolbar = React.forwardRef(
-  (
-    { className, ...props }: PropsWithChildren<BaseProps>,
-    ref: Ref<HTMLDivElement>,
-  ) => (
-    <Menu
-      {...props}
-      ref={ref}
-      className={className}
-      sx={{
-        position: 'relative',
-        marginBottom: '20px',
-      }}
-    />
-  ),
-);
-
-export const Element = ({
-  attributes,
-  children,
-  element,
-}: {
-  attributes: any;
-  children: any;
-  element: any;
-}) => {
-  switch (element.type) {
-    case 'block-quote':
-      return <blockquote {...attributes}>{children}</blockquote>;
-    case 'bulleted-list':
-      return <ul {...attributes}>{children}</ul>;
-    case 'heading-one':
-      return <h1 {...attributes}>{children}</h1>;
-    case 'heading-two':
-      return <h2 {...attributes}>{children}</h2>;
-    case 'list-item':
-      return <li {...attributes}>{children}</li>;
-    case 'numbered-list':
-      return <ol {...attributes}>{children}</ol>;
-    default:
-      return <p {...attributes}>{children}</p>;
-  }
-};
-
-export const Leaf = ({
-  attributes,
-  children,
-  leaf,
-}: {
-  attributes: any;
-  children: any;
-  leaf: any;
-}) => {
-  if (leaf.bold) {
-    children = <strong>{children}</strong>;
-  }
-
-  if (leaf.code) {
-    children = <code>{children}</code>;
-  }
-
-  if (leaf.italic) {
-    children = <em>{children}</em>;
-  }
-
-  if (leaf.underline) {
-    children = <u>{children}</u>;
-  }
-
-  return <span {...attributes}>{children}</span>;
-};
