+++ src/components/index.ts
@@ -4,4 +4,4 @@ export * from './LoaderIcon/LoaderIcon';
 export * from './LoaderButton/LoaderButton';
 export * from './Skeleton/Skeleton';
 export * from './Table/Table';
-export * from './WYSIWYGEditor/WYSIWYGEditor';
+export * from './Editor/Editor';
